rag_service/requirement.txt
fastapi>=0.116.0
uvicorn>=0.30.0
faiss-cpu>=1.8.0
numpy>=1.26.0
openai>=1.40.0
python-multipart>=0.0.9
pypdf>=4.3.0
python-docx>=1.1.0
Pillow>=10.4.0